pyproject.toml
[project]
name = "quartz-server-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.11",
    "alembic>=1.14.1",
    "fastapi[standard]>=0.115.6",
    "google-generativeai>=0.8.4",
    "grpcio==1.67.1",
    "grpcio-status==1.67.1",
    "passlib[argon2]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.4",
    "pydantic-core>=2.27.2",
    "pydantic-settings>=2.7.1",
    "pyjwt>=2.10.1",
    "sqlalchemy[aiosqlite]>=2.0.37",
    "sqlmodel>=0.0.22",
    "typing>=3.10.0.0",
]
